+++ skcriteria/preprocessing/weighters.py
@@ -30,7 +30,7 @@ import scipy.stats
 
 
 from ._preprocessing_base import SKCTransformerABC
-from .distance import cenit_distance
+from .scalers import cenit_distance
 from ..core import Objective
 from ..utils import doc_inherit